Update to 0.0.7. Blockly is not enabled by default.  Check the console log to learn how to enable it. | Moved Blockly functions to blockly.js.
