--- editincode
@@ -1,5 +1,0 @@
-#! /bin/bash
-
-code .
-
-exit 0
