+++ package.json
@@ -1,11 +1,12 @@
 {
-  "name": "QEdit",
-  "productName": "QEdit",
-  "version": "0.0.6",
+  "name": "QEdit-KV_Hack",
+  "productName": "QEdit-KV_Hack",
+  "version": "0.0.7",
   "description": "Editor for Quest 6 (hacked by KV)",
   "main": "src/index.js",
   "scripts": {
     "start": "electron-forge start",
+    "dev": "electron-forge start  --inspect-electron",
     "package": "electron-forge package",
     "make": "electron-forge make",
     "publish": "electron-forge publish",
@@ -13,7 +14,7 @@
     "test": "jest"
   },
   "keywords": [],
-  "author": "AndyJ <the_pix@hotmail.com>",
+  "authors": "AndyJ <the_pix@hotmail.com> & KV",
   "license": "MIT",
   "jest": {
     "testEnvironment": "jest-environment-jsdom-fourteen"
@@ -23,7 +24,7 @@
     "forge": {
       "make_targets": {
         "win32": [
-          "squirrel"
+
         ],
         "darwin": [
           "zip"
