+++ src/index.html
@@ -16,24 +16,6 @@
 <head>
 <body style="overflow: hidden; margin: auto;" >
   <div id="App"></div>
-  <hr/>
-  <textarea id="textarea" placeholder="//Your code will appear here.  Copy it from here and paste it." style="width:500px;height:50px;margin-left:20%;"></textarea>
-  <div id="blocklyDiv" style="height: 600px; width: 800px; margin-left:20%;"></div>
-  
-  <xml id="toolbox" style="display: none">
-      <block type="controls_if"></block>
-      <block type="controls_repeat_ext"></block>
-      <block type="logic_compare"></block>
-      <block type="math_number"></block>
-      <block type="math_arithmetic"></block>
-      <block type="text"></block>
-      <block type="text_print"></block>
-      <block type="logic_boolean"></block>
-      <block type="msg"></block>
-      <block type="return"></block>
-      <block type="variables_get"></block>
-      <block type="variables_set"></block>
-    </xml>
 </body>
 
 <script type="text/jsx">
